Upgrade to NEST 7.x

+++ elasticsearch-nest-webapi-angularjs/Models/SearchResult.cs
@@ -10,7 +10,7 @@ namespace elasticsearch_nest_webapi_angularjs.Models
 
         public IEnumerable<T> Results { get; set; }
 
-        public int ElapsedMilliseconds { get; set; }
+        public long ElapsedMilliseconds { get; set; }
 
         public Dictionary<string, long> AggregationsByTags { get; set; }
     }
